Fix bug causing the Ethernet not to work
The Ethernet was not working. The HIETH MII/RMII mode options in the Linux
configuration needed changing to 1 (RMII) and a bug in the net_rmii_mode function
of the pinmux_hi3518.sh startup script needed fixing, which was setting the
muxctrl_reg28 register to the wrong function (EPHY_CLK instead of GPIO1_3).